chore: Move graph-related hooks files.
