+++ src/plugins/graph/axis/components/axis-or-legend-attribute-menu.tsx
@@ -4,9 +4,9 @@ import t from "../../utilities/translation/translate";
 import {GraphPlace} from "../../axis-graph-shared";
 import { graphPlaceToAttrRole } from "../../graph-types";
 import { useDataConfigurationContext } from "../../hooks/use-data-configuration-context";
-import { useDataSetContext } from "../../../../hooks/use-data-set-context";
-import { useOutsidePointerDown } from "../../../../hooks/use-outside-pointer-down";
-import { useOverlayBounds } from "../../../../hooks/use-overlay-bounds";
+import { useDataSetContext } from "../../hooks/use-data-set-context";
+import { useOutsidePointerDown } from "../../hooks/use-outside-pointer-down";
+import { useOverlayBounds } from "../../hooks/use-overlay-bounds";
 import { AttributeType } from "../../../../models/data/attribute";
 
 interface IProps {
