+++ src/plugins/graph/components/casedots.tsx
@@ -7,8 +7,8 @@ import {ICase} from "../../../models/data/data-set-types";
 import {isAddCasesAction} from "../../../models/data/data-set-actions";
 import {useDragHandlers, usePlotResponders} from "../hooks/use-plot";
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context";
-import {useDataSetContext} from "../../../hooks/use-data-set-context";
-import {useInstanceIdContext} from "../../../hooks/use-instance-id-context";
+import {useDataSetContext} from "../hooks/use-data-set-context";
+import {useInstanceIdContext} from "../hooks/use-instance-id-context";
 import {useGraphLayoutContext} from "../models/graph-layout";
 import {handleClickOnDot, setPointSelection} from "../utilities/graph-utils";
 import {defaultSelectedStroke, defaultSelectedStrokeWidth, defaultStrokeWidth} from "../../../utilities/color-utils";
