+++ src/plugins/graph/components/droppable-add-attribute.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import {clsx} from "clsx";
 import {Active, useDroppable} from "@dnd-kit/core";
-import {useDropHintString} from "../../../hooks/use-drop-hint-string";
-import {getDragAttributeId, useDropHandler} from "../../../hooks/use-drag-drop";
+import {useDropHintString} from "../hooks/use-drop-hint-string";
+import {getDragAttributeId, useDropHandler} from "../hooks/use-drag-drop";
 import {DropHint} from "./drop-hint";
 import {graphPlaceToAttrRole, PlotType} from "../graph-types";
 import {GraphPlace} from "../axis-graph-shared";
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context";
-import {useInstanceIdContext} from "../../../hooks/use-instance-id-context";
+import {useInstanceIdContext} from "../hooks/use-instance-id-context";
 
 interface IAddAttributeProps {
   place: GraphPlace
