+++ src/plugins/graph/components/droppable-plot.tsx
@@ -1,8 +1,8 @@
 import {Active} from "@dnd-kit/core";
 import React, {memo} from "react";
-import {getDragAttributeId, useDropHandler} from "../../../hooks/use-drag-drop";
-import {useDropHintString} from "../../../hooks/use-drop-hint-string";
-import {useInstanceIdContext} from "../../../hooks/use-instance-id-context";
+import {getDragAttributeId, useDropHandler} from "../hooks/use-drag-drop";
+import {useDropHintString} from "../hooks/use-drop-hint-string";
+import {useInstanceIdContext} from "../hooks/use-instance-id-context";
 import {DroppableSvg} from "./droppable-svg";
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context";
 import {GraphPlace} from "../axis-graph-shared";
