+++ src/plugins/graph/components/graph-component.tsx
@@ -2,11 +2,11 @@ import {useDroppable} from '@dnd-kit/core';
 import {observer} from "mobx-react-lite";
 import React, {useEffect, useMemo, useRef} from "react";
 import {useResizeDetector} from "react-resize-detector";
-import {useDataSet} from '../../../hooks/use-data-set';
-import {DataSetContext} from '../../../hooks/use-data-set-context';
+import {useDataSet} from '../hooks/use-data-set';
+import {DataSetContext} from '../hooks/use-data-set-context';
 import {useGraphController} from "../hooks/use-graph-controller";
 import {useInitGraphLayout} from '../hooks/use-init-graph-layout';
-import {InstanceIdContext, useNextInstanceId} from "../../../hooks/use-instance-id-context";
+import {InstanceIdContext, useNextInstanceId} from "../hooks/use-instance-id-context";
 import {AxisLayoutContext} from "../axis/models/axis-layout-context";
 import {GraphController} from "../models/graph-controller";
 import {GraphLayoutContext} from "../models/graph-layout";
