+++ src/plugins/graph/components/graph.tsx
@@ -15,14 +15,14 @@ import {CaseDots} from "./casedots";
 import {ChartDots} from "./chartdots";
 import {Marquee} from "./marquee";
 import {DataConfigurationContext} from "../hooks/use-data-configuration-context";
-import {useDataSetContext} from "../../../hooks/use-data-set-context";
+import {useDataSetContext} from "../hooks/use-data-set-context";
 import {useGraphModel} from "../hooks/use-graph-model";
 import {setNiceDomain, startAnimation} from "../utilities/graph-utils";
 import {IAxisModel} from "../axis/models/axis-model";
 import {GraphPlace} from "../axis-graph-shared";
 import {useGraphLayoutContext} from "../models/graph-layout";
 import {isSetAttributeIDAction, useGraphModelContext} from "../models/graph-model";
-import {useInstanceIdContext} from "../../../hooks/use-instance-id-context";
+import {useInstanceIdContext} from "../hooks/use-instance-id-context";
 import {MarqueeState} from "../models/marquee-state";
 import {Legend} from "./legend/legend";
 import {AttributeType} from "../../../models/data/attribute";
