+++ src/plugins/graph/components/legend/legend.tsx
@@ -6,9 +6,9 @@ import {AttributeLabel} from "../attribute-label";
 import {CategoricalLegend} from "./categorical-legend";
 import {NumericLegend} from "./numeric-legend";
 import {DroppableSvg} from "../droppable-svg";
-import {useInstanceIdContext} from "../../../../hooks/use-instance-id-context";
-import {getDragAttributeId, useDropHandler} from "../../../../hooks/use-drag-drop";
-import {useDropHintString} from "../../../../hooks/use-drop-hint-string";
+import {useInstanceIdContext} from "../../hooks/use-instance-id-context";
+import {getDragAttributeId, useDropHandler} from "../../hooks/use-drag-drop";
+import {useDropHintString} from "../../hooks/use-drop-hint-string";
 import { AttributeType } from "../../../../models/data/attribute";
 import {GraphAttrRole} from "../../graph-types";
 import {GraphPlace} from "../../axis-graph-shared";
