+++ src/plugins/graph/hooks/use-overlay-bounds.ts
